+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2024-12-02
+
 ### Added
 
 - Checkbox to accept untrusted SSL certificates #72
@@ -338,7 +340,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Intiial release
 
-[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.11.0...HEAD
+[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/tmaegel/ntodotxt/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/tmaegel/ntodotxt/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/tmaegel/ntodotxt/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/tmaegel/ntodotxt/compare/v0.9.1...v0.10.0